+++ app/src/main/java/com/example/testapplication/model/Clouds.kt
@@ -1,5 +1,5 @@
 package com.example.testapplication.model
 
 data class Clouds(
-    val all: Int
+    val all: Double
 )
